+++ TidalPy/radial_solver/__init__.py
@@ -1,1 +1,2 @@
 from .numerical import radial_solver, radial_solver_numba
+from .love import find_love